Testing improvements (#2)
* Reworked TestHarness and clean it up. Enabled the test inclusion/exclusion options.

* Primarily getting acceptedValues to work with regular expressions along with primitives

* Added value masking for testing values so when they are pushed to Kafka, PII can be masked out.

* Minor adjustment to ensure if value for regexp doesn't match

* Moved test config storage to a cache and changed how the keys are created.

This is to enhance testing speed so each hit of a method doesn't require a parsing routine. Currently set ttl to 5 minutes and a limit of 5000 entries which is all managed for the developer.

* Little bit of code clean up.

* Commented out default settings for AppCache as empty is a valid option for node-cache

* Fix for DSL parsing of object test config.

Fixed parsing of regexp for value checking.
Fix for isOr flag so if more than one object config is in the required array, they will all be or
If multiple properties are defined within one required config, they will be and.

* Changes to make object testing functional with simple type checking and value validation via array or regexp.

This also supports and-ing and or-ing between properties.

* Changes to return the method name if available otherwise returns path to js file.

* Fix for analyzing untested params and slight code clean up

* Additional testing functionality

Added array type checking if a subtype is passed in config, i.e. array<number>
Added internal object value masking if value is a string

* Got output testing to work with template strings and complex objects.